Extracts out an rtf.rs
